tweet_to_slack/requirements.txt
slackclient
nest_asyncio
psycopg2-binary
tweepy ==4.12.1
pymongo ==4.3.3
pandas ==1.5.2
sqlalchemy ==1.4.45

